PBC/requirements.txt
absl-py==0.7.1
alembic==1.2.1
amqp==2.5.0
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
blinker==1.4
brotlipy==0.7.0
captcha==0.3
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
click==6.7
colorama==0.4.1
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
Django==2.2.4
django-cors-headers==3.1.0
docopt==0.6.2
fake-useragent==0.1.11
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
fonttools==3.44.0
gast==0.2.2
gerapy==0.8.5
gevent==1.4.0
google-pasta==0.1.7
greenlet==0.4.15
grequests==0.4.0
grpcio==1.22.0
h11==0.7.0
h2==3.1.1
h5py==2.9.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.17
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
Js2Py==0.66
kaitaistruct==0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.3
lazy-object-proxy==1.3.1
ldap3==2.5.2
librabbitmq==2.0.0
lxml==4.3.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mitmproxy==4.0.4
msgpack==0.6.1
numpy==1.16.3
opencv-python==4.1.1.26
parsel==1.5.1
passlib==1.7.1
Pillow==6.0.0
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
pydivert==2.1.0
pyee==6.0.0
PyExecJS==1.5.1
PyHamcrest==1.9.0
pyjsparser==2.7.1
pylint==2.3.1
pymongo==3.7.2
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
pypiwin32==223
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.8.0
python-editor==1.0.4
python-scrapyd-api==2.1.2
pytils==0.3
pytz==2019.1
pywin32==224
queuelib==1.5.0
readability-lxml==0.7.1
redis==3.2.1
requests==2.21.0
ruamel.yaml==0.15.100
Scrapy==1.6.0
scrapy-redis==0.6.8
scrapy-redis-bloomfilter==0.7.0
scrapy-splash==0.7.2
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sortedcontainers==2.0.5
soupsieve==1.9
SQLAlchemy==1.3.8
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tools==0.1.9
tornado==5.1.1
tqdm==4.33.0
Twisted==19.7.0
typed-ast==1.3.1
tzlocal==2.0.0
urllib3==1.25.6
urwid==2.0.1
vine==1.3.0
virtualenv==16.7.2
virtualenvwrapper-win==1.2.5
w3lib==1.20.0
websockets==8.0.2
Werkzeug==0.15.5
wordcloud==1.5.0
wrapt==1.11.1
wsproto==0.11.0
WTForms==2.2.1
zope.interface==4.6.0
